Move fetch function `getPlaces()` out of map.js and into own file
